fix empty results when result set is too large to cache
`cache_get_data` may return `false` when the cached data was too large
for the configured cache backend to store the value. This results in the
sesarch results becoming empty when users refresh or move to the next
page.

This fixes this by checking if the result of `cache_get_data` is an
array before attempting to use it.